Crate google-fusiontables2

This documentation was generated from fusiontables crate version 0.0.1+20141113, where 20141113 is the exact revision of the fusiontables:v2 schema built by the mako code generator v0.0.1.

Everything else about the fusiontables v2 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

column

delete, get, insert, list, patch and update

query

sql and sql get

style

delete, get, insert, list, patch and update

table

copy, delete, get, import rows, import table, insert, list, patch, replace rows and update

task

delete, get and list

template

delete, get, insert, list, patch and update

Upload supported by ...

replace rows table

import rows table

import table table

Download supported by ...

sql get query

sql query

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.table().import_table(...).doit()
let r = hub.table().patch(...).doit()
let r = hub.table().get(...).doit()
let r = hub.table().copy(...).doit()
let r = hub.table().insert(...).doit()
let r = hub.table().update(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-fusiontables2 = "*"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "google-fusiontables2" as fusiontables2;
use fusiontables2::Result;
use std::fs;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use fusiontables2::Fusiontables;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Fusiontables::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `upload_resumable(...)`.
// Values shown here are possibly random and not representative !
let result = hub.table().import_table("name")
             .encoding("et")
             .delimiter("dolores")
             .upload_resumable(fs::File::open("file.ext").unwrap(), "application/octet-stream".parse().unwrap());
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Auth: Missing API Key - used if there are no scopes"),
    Result::MissingToken => println!("OAuth2: Missing Token"),
    Result::UploadSizeLimitExceeded(size, max_size) => println!("Upload size too big: {} of {}", size, max_size),
    Result::Failure(_) => println!("General Failure (hyper::client::Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("You added custom parameter which is part of builder: {:?}", clashed_field),
    Result::JsonDecodeError(err) => println!("Couldn't understand server reply - maybe API needs update: {:?}", err),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downlods

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

Bucket
Specifies the minimum and maximum values, the color, opacity, icon and weight of a bucket within a StyleSetting.

Column
Specifies the details of a column in a table.

ColumnBaseColumn
Identifier of the base column. If present, this column is derived from the specified base column.

ColumnDeleteCall
Deletes the column.

ColumnGetCall
Retrieves a specific column by its id.

ColumnInsertCall
Adds a new column to the table.

ColumnList
Represents a list of columns in a table.

ColumnListCall
Retrieves a list of columns.

ColumnMethods
A builder providing access to all methods supported on column resources. It is not used directly, but through the Fusiontables hub.

ColumnPatchCall
Updates the name or type of an existing column. This method supports patch semantics.

ColumnUpdateCall
Updates the name or type of an existing column.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

Fusiontables
Central instance to access all Fusiontables related resource activities

Import
Represents an import request.

JsonServerError
A utility type which can decode a server response that indicates error

LineStyle
Represents a LineStyle within a StyleSetting

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

PointStyle
Represents a PointStyle within a StyleSetting

PolygonStyle
Represents a PolygonStyle within a StyleSetting

QueryMethods
A builder providing access to all methods supported on query resources. It is not used directly, but through the Fusiontables hub.

QuerySqlCall
Executes an SQL SELECT/INSERT/UPDATE/DELETE/SHOW/DESCRIBE/CREATE statement.

QuerySqlGetCall
Executes an SQL SELECT/SHOW/DESCRIBE statement.

Sqlresponse
Represents a response to an sql statement.

StyleDeleteCall
Deletes a style.

StyleFunction
Represents a StyleFunction within a StyleSetting

StyleFunctionGradient
Gradient function that interpolates a range of colors based on column value.

StyleFunctionGradientColors
Array with two or more colors.

StyleGetCall
Gets a specific style.

StyleInsertCall
Adds a new style for the table.

StyleListCall
Retrieves a list of styles.

StyleMethods
A builder providing access to all methods supported on style resources. It is not used directly, but through the Fusiontables hub.

StylePatchCall
Updates an existing style. This method supports patch semantics.

StyleSetting
Represents a complete StyleSettings object. The primary key is a combination of the tableId and a styleId.

StyleSettingList
Represents a list of styles for a given table.

StyleUpdateCall
Updates an existing style.

Table
Represents a table.

TableCopyCall
Copies a table.

TableDeleteCall
Deletes a table.

TableGetCall
Retrieves a specific table by its id.

TableImportRowCall
Import more rows into a table.

TableImportTableCall
Import a new table.

TableInsertCall
Creates a new table.

TableList
Represents a list of tables.

TableListCall
Retrieves a list of tables a user owns.

TableMethods
A builder providing access to all methods supported on table resources. It is not used directly, but through the Fusiontables hub.

TablePatchCall
Updates an existing table. Unless explicitly requested, only the name, description, and attribution will be updated. This method supports patch semantics.

TableReplaceRowCall
Replaces rows of an existing table. Current rows remain visible until all replacement rows are ready.

TableUpdateCall
Updates an existing table. Unless explicitly requested, only the name, description, and attribution will be updated.

Task
A background task on a table, initiated for time and or resource consuming operations such as a column type change or delete all rows operation.

TaskDeleteCall
Deletes the task, unless already started.

TaskGetCall
Retrieves a specific task by its id.

TaskList
Represents a list of tasks in a table.

TaskListCall
Retrieves a list of tasks.

TaskMethods
A builder providing access to all methods supported on task resources. It is not used directly, but through the Fusiontables hub.

Template
Represents the contents of InfoWindow templates.

TemplateDeleteCall
Deletes a template

TemplateGetCall
Retrieves a specific template by its id

TemplateInsertCall
Creates a new template for the table.

TemplateList
Represents a list of templates for a given table.

TemplateListCall
Retrieves a list of templates.

TemplateMethods
A builder providing access to all methods supported on template resources. It is not used directly, but through the Fusiontables hub.

TemplatePatchCall
Updates an existing template. This method supports patch semantics.

TemplateUpdateCall
Updates an existing template

Enums

Result
A universal result type used as return for all action method results.

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResourceMethodsBuilder
Identifies types for building methods of a particular resource type

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string