Struct google_fusiontables2::Bucket

pub struct Bucket {
    pub opacity: Option<f64>,
    pub weight: Option<i32>,
    pub min: Option<f64>,
    pub color: Option<String>,
    pub max: Option<f64>,
    pub icon: Option<String>,
}

Specifies the minimum and maximum values, the color, opacity, icon and weight of a bucket within a StyleSetting.

This type is not used in any activity, and only used as part of another schema.

Fields

opacity

Opacity of the color: 0.0 (transparent) to 1.0 (opaque).

weight

Width of a line (in pixels).

min

Minimum value in the selected column for a row to be styled according to the bucket color, opacity, icon, or weight.

color

Color of line or the interior of a polygon in #RRGGBB format.

max

Maximum value in the selected column for a row to be styled according to the bucket color, opacity, icon, or weight.

icon

Icon name used for a point.