Struct google_fusiontables2::PolygonStyle

pub struct PolygonStyle {
    pub stroke_color_styler: Option<StyleFunction>,
    pub stroke_weight: Option<i32>,
    pub stroke_opacity: Option<f64>,
    pub stroke_weight_styler: Option<StyleFunction>,
    pub fill_color_styler: Option<StyleFunction>,
    pub fill_color: Option<String>,
    pub stroke_color: Option<String>,
    pub fill_opacity: Option<f64>,
}

Represents a PolygonStyle within a StyleSetting

This type is not used in any activity, and only used as part of another schema.

Fields

stroke_color_styler

Column-value, gradient or buckets styler that is used to determine the border color and opacity.

stroke_weight

Width of the polyon border in pixels.

stroke_opacity

Opacity of the polygon border: 0.0 (transparent) to 1.0 (opaque).

stroke_weight_styler

Column-value or bucket styler that is used to determine the width of the polygon border.

fill_color_styler

Column-value, gradient, or bucket styler that is used to determine the interior color and opacity of the polygon.

fill_color

Color of the interior of the polygon in #RRGGBB format.

stroke_color

Color of the polygon border in #RRGGBB format.

fill_opacity

Opacity of the interior of the polygon: 0.0 (transparent) to 1.0 (opaque).