Struct google_fusiontables2::TableMethods

pub struct TableMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on table resources. It is not used directly, but through the Fusiontables hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_fusiontables2 as fusiontables2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use fusiontables2::Fusiontables;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Fusiontables::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `copy(...)`, `delete(...)`, `get(...)`, `import_rows(...)`, `import_table(...)`, `insert(...)`, `list(...)`, `patch(...)`, `replace_rows(...)` and `update(...)`
// to build up your call.
let rb = hub.table();

Methods

impl<'a, C, A> TableMethods<'a, C, A>

fn patch(&self, request: Table, table_id: &str) -> TablePatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an existing table. Unless explicitly requested, only the name, description, and attribution will be updated. This method supports patch semantics.

Arguments

request - No description provided.

tableId - ID of the table that is being updated.

fn replace_rows(&self, table_id: &str) -> TableReplaceRowCall<'a, C, A>

Create a builder to help you perform the following task:

Replaces rows of an existing table. Current rows remain visible until all replacement rows are ready.

Arguments

tableId - Table whose rows will be replaced.

fn update(&self, request: Table, table_id: &str) -> TableUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an existing table. Unless explicitly requested, only the name, description, and attribution will be updated.

Arguments

request - No description provided.

tableId - ID of the table that is being updated.

fn delete(&self, table_id: &str) -> TableDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a table.

Arguments

tableId - ID of the table to be deleted.

fn insert(&self, request: Table) -> TableInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new table.

Arguments

request - No description provided.

fn get(&self, table_id: &str) -> TableGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a specific table by its ID.

Arguments

tableId - Identifier for the table being requested.

fn list(&self) -> TableListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a list of tables a user owns.

fn import_rows(&self, table_id: &str) -> TableImportRowCall<'a, C, A>

Create a builder to help you perform the following task:

Imports more rows into a table.

Arguments

tableId - The table into which new rows are being imported.

fn import_table(&self, name: &str) -> TableImportTableCall<'a, C, A>

Create a builder to help you perform the following task:

Imports a new table.

Arguments

name - The name to be assigned to the new table.

fn copy(&self, table_id: &str) -> TableCopyCall<'a, C, A>

Create a builder to help you perform the following task:

Copies a table.

Arguments

tableId - ID of the table that is being copied.