Struct google_fusiontables2::Task

pub struct Task {
    pub started: Option<bool>,
    pub progress: Option<String>,
    pub kind: Option<String>,
    pub type_: Option<String>,
    pub task_id: Option<String>,
}

A background task on a table, initiated for time- or resource-consuming operations such as changing column types or deleting all rows.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get task (response)

replace rows table (response)

Fields
started: Option<bool>

false while the table is busy with some other task. true if this background task is currently running.

progress: Option<String>

Task percentage completion.

kind: Option<String>

Type of the resource. This is always "fusiontables#task".

type_: Option<String>

Type of background task.

task_id: Option<String>

Identifier for the task.