Struct google_fusiontables2::Column

pub struct Column {
    pub graph_predicate: Option<String>,
    pub kind: Option<String>,
    pub valid_values: Option<Vec<String>>,
    pub description: Option<String>,
    pub base_column: Option<ColumnBaseColumn>,
    pub column_id: Option<i32>,
    pub column_properties_json: Option<String>,
    pub format_pattern: Option<String>,
    pub column_json_schema: Option<String>,
    pub type_: Option<String>,
    pub validate_data: Option<bool>,
    pub name: Option<String>,
}

Specifies the details of a column in a table.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get column (response)

update column (request|response)

insert column (request|response)

patch column (request|response)

Fields

graph_predicate

Column graph predicate. Used to map table to graph data model (subject,predicate,object) See W3C Graph-based Data Model.

kind

The kind of item this is. For a column, this is always fusiontables#column.

valid_values

List of valid values used to validate data and supply a drop-down list of values in the web application.

description

Column description.

base_column

Identifier of the base column. If present, this column is derived from the specified base column.

column_id

Identifier for the column.

column_properties_json

JSON object containing custom column properties.

format_pattern

Format pattern. Acceptable values are DT_DATE_MEDIUMe.g Dec 24, 2008 DT_DATE_SHORTfor example 12/24/08 DT_DATE_TIME_MEDIUMfor example Dec 24, 2008 8:30:45 PM DT_DATE_TIME_SHORTfor example 12/24/08 8:30 PM DT_DAY_MONTH_2_DIGIT_YEARfor example 24/12/08 DT_DAY_MONTH_2_DIGIT_YEAR_TIMEfor example 24/12/08 20:30 DT_DAY_MONTH_2_DIGIT_YEAR_TIME_MERIDIANfor example 24/12/08 8:30 PM DT_DAY_MONTH_4_DIGIT_YEARfor example 24/12/2008 DT_DAY_MONTH_4_DIGIT_YEAR_TIMEfor example 24/12/2008 20:30 DT_DAY_MONTH_4_DIGIT_YEAR_TIME_MERIDIANfor example 24/12/2008 8:30 PM DT_ISO_YEAR_MONTH_DAYfor example 2008-12-24 DT_ISO_YEAR_MONTH_DAY_TIMEfor example 2008-12-24 20:30:45 DT_MONTH_DAY_4_DIGIT_YEARfor example 12/24/2008 DT_TIME_LONGfor example 8:30:45 PM UTC-6 DT_TIME_MEDIUMfor example 8:30:45 PM DT_TIME_SHORTfor example 8:30 PM DT_YEAR_ONLYfor example 2008 HIGHLIGHT_UNTYPED_CELLSHighlight cell data that does not match the data type NONENo formatting (default) NUMBER_CURRENCYfor example $1234.56 NUMBER_DEFAULTfor example 1,234.56 NUMBER_INTEGERfor example 1235 NUMBER_NO_SEPARATORfor example 1234.56 NUMBER_PERCENTfor example 123,456% NUMBER_SCIENTIFICfor example 1E3 STRING_EIGHT_LINE_IMAGEDisplays thumbnail images as tall as eight lines of text STRING_FOUR_LINE_IMAGEDisplays thumbnail images as tall as four lines of text STRING_JSON_TEXTAllows JSON editing of text in UI STRING_LINKTreats cell as a link (must start with http:// or https://) STRING_ONE_LINE_IMAGEDisplays thumbnail images as tall as one line of text STRING_VIDEO_OR_MAPDisplay a video or map thumbnail

column_json_schema

JSON schema for interpreting JSON in this column.

type_

Type of the column.

validate_data

If true, data entered via the web application is validated.

name

Name of the column.